Makes wallpaper change occur first.
The wallpaper change would occur after pywal
calculated the colors and made all the themes.
This was because the wallpaper needed to be set
manually with gsettings in order for cron to be
able to change the wallpaper image. The image
was taken from pywal's cache. Now 'sort -R' is used
 to randomly pick an image from the wallpapers folder.
Thus the walpaper can be set with gsettings first, and
then pywal can be allowed to work.