Changes installtion folder to be general
Previously the git repository needed to be in a specific
place in order for the pieces to be accessible. Now the pieces
folder is sym linked to the home folder. Thus, the git repository
can be anywhere, the installation file just needs to be run.

Additionally, all the pieces were moved into a single folder.