Adds .tmux as tmux configuration
It's a very cool tmux configuration that is a seperate git
repository. It was thus added as a submodule. An installtion
script was made and a .tmux.conf.local as saved outside of the
repository. This is because that is what one is suppossed to
edit as a user of this configuraiton.