Adds http server of current background for "tfac"
"tfac" is Tab For a Cause. It allows you to specify
a custom background only through a url. Thus the
http server of the current background. It is
serving files in /tmp/background_server. In that
directory is a file called current_background.
This file is a copy of the current background.
This script checks  if http-server is already running
before starting the server.